+++ OneBlink.SDK/models/GetSubmissionPdfRequest.cs
@@ -29,5 +29,9 @@ namespace OneBlink.SDK.Model
         {
             get; set;
         }
+        public bool? includeExternalIdInPdf
+        {
+            get; set;
+        }
     }
 }